выполнено задание 5 - деление на модули
